+++ tests/spot/portfoliomargin/PortfolioMarginCollateralRateTest.php
@@ -1,0 +1,30 @@
+<?php
+
+use Binance\Tests\BaseTestCase;
+use Aeris\GuzzleHttpMock\Expect;
+
+class PortfolioMarginCollateralRateTest extends BaseTestCase
+{
+    public function setUp(): void
+    {
+        parent::setUp();
+    }
+
+    public function testPortfolioMarginCollateralRate()
+    {
+        $this->httpMock
+            ->shouldReceiveRequest()
+            ->withUrl(new Expect\Equals('/sapi/v1/portfolio/collateralRate'))
+            ->withMethod('GET')
+            ->andRespondWithJson($this->data, $statusCode = 200);
+
+        $response = $this->spotClient->portfolioMarginCollateralRate();
+
+        $this->assertEquals($response, $this->data);
+    }
+
+    public function tearDown(): void
+    {
+        parent::tearDown();
+    }
+}